# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MLSystem2 Training UI API",
    "description": "Public FastAPI contract for the MLSystem2 training site.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/training-templates": {
      "get": {
        "summary": "Get Training Templates",
        "operationId": "get_training_templates_api_v1_training_templates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingTemplateListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Post Training Template",
        "operationId": "post_training_template_api_v1_training_templates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingTemplateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingTemplate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/training-templates/by-id/{template_id}": {
      "put": {
        "summary": "Put Training Template By Id",
        "operationId": "put_training_template_by_id_api_v1_training_templates_by_id__template_id__put",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingTemplateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingTemplate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Training Template By Id",
        "operationId": "delete_training_template_by_id_api_v1_training_templates_by_id__template_id__delete",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingTemplate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/training-templates/by-id/{template_id}/apply-field-to-all": {
      "put": {
        "summary": "Put Training Template Field To All",
        "operationId": "put_training_template_field_to_all_api_v1_training_templates_by_id__template_id__apply_field_to_all_put",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingTemplateApplyField"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingTemplateListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/training-templates/{architecture}": {
      "get": {
        "summary": "Get Training Template",
        "operationId": "get_training_template_api_v1_training_templates__architecture__get",
        "parameters": [
          {
            "name": "architecture",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Architecture"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingTemplate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Put Training Template",
        "operationId": "put_training_template_api_v1_training_templates__architecture__put",
        "parameters": [
          {
            "name": "architecture",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Architecture"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingTemplateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingTemplate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inference-templates": {
      "get": {
        "summary": "Get Inference Templates",
        "operationId": "get_inference_templates_api_v1_inference_templates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceTemplateListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Post Inference Template",
        "operationId": "post_inference_template_api_v1_inference_templates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingTemplateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceTemplate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inference-templates/by-id/{template_id}": {
      "put": {
        "summary": "Put Inference Template By Id",
        "operationId": "put_inference_template_by_id_api_v1_inference_templates_by_id__template_id__put",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingTemplateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceTemplate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Inference Template By Id",
        "operationId": "delete_inference_template_by_id_api_v1_inference_templates_by_id__template_id__delete",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceTemplate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inference-templates/by-id/{template_id}/apply-field-to-all": {
      "put": {
        "summary": "Put Inference Template Field To All",
        "operationId": "put_inference_template_field_to_all_api_v1_inference_templates_by_id__template_id__apply_field_to_all_put",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingTemplateApplyField"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceTemplateListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inference-templates/{architecture}": {
      "get": {
        "summary": "Get Inference Template",
        "operationId": "get_inference_template_api_v1_inference_templates__architecture__get",
        "parameters": [
          {
            "name": "architecture",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Architecture"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceTemplate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Put Inference Template",
        "operationId": "put_inference_template_api_v1_inference_templates__architecture__put",
        "parameters": [
          {
            "name": "architecture",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Architecture"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingTemplateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceTemplate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigField": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value_type": {
            "type": "string",
            "title": "Value Type"
          },
          "tooltip": {
            "type": "string",
            "title": "Tooltip"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": true
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "min_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Value"
          },
          "max_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Value"
          },
          "recommended_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Range"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "label",
          "value_type",
          "tooltip"
        ],
        "title": "ConfigField"
      },
      "ConfigSchema": {
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ConfigField"
            },
            "type": "array",
            "title": "Fields"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fields"
        ],
        "title": "ConfigSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InferenceTemplate": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "architecture": {
            "type": "string",
            "title": "Architecture"
          },
          "dataset_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Key"
          },
          "dataset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Name"
          },
          "parent_template_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Template Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "config_schema": {
            "$ref": "#/components/schemas/ConfigSchema"
          },
          "default_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Default Config"
          },
          "source": {
            "$ref": "#/components/schemas/TemplateSource"
          },
          "source_mlflow_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Mlflow Run Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "architecture",
          "display_name",
          "config_schema",
          "default_config",
          "source",
          "is_active",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "InferenceTemplate"
      },
      "InferenceTemplateListResponse": {
        "properties": {
          "templates": {
            "items": {
              "$ref": "#/components/schemas/InferenceTemplate"
            },
            "type": "array",
            "title": "Templates"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "templates"
        ],
        "title": "InferenceTemplateListResponse"
      },
...
      "TemplateSource": {
        "type": "string",
        "enum": [
          "hpo_best",
          "analogy",
          "manual"
        ],
        "title": "TemplateSource"
      },
...
      "TrainingTemplate": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "architecture": {
            "type": "string",
            "title": "Architecture"
          },
          "dataset_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Key"
          },
          "dataset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Name"
          },
          "parent_template_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Template Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "config_schema": {
            "$ref": "#/components/schemas/ConfigSchema"
          },
          "default_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Default Config"
          },
          "source": {
            "$ref": "#/components/schemas/TemplateSource"
          },
          "source_mlflow_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Mlflow Run Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "architecture",
          "display_name",
          "config_schema",
          "default_config",
          "source",
          "is_active",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "TrainingTemplate"
      },
      "TrainingTemplateApplyField": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "title": "Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "TrainingTemplateApplyField"
      },
      "TrainingTemplateCreate": {
        "properties": {
          "architecture": {
            "type": "string",
            "title": "Architecture"
          },
          "dataset_key": {
            "type": "string",
            "title": "Dataset Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "architecture",
          "dataset_key"
        ],
        "title": "TrainingTemplateCreate"
      },
      "TrainingTemplateListResponse": {
        "properties": {
          "templates": {
            "items": {
              "$ref": "#/components/schemas/TrainingTemplate"
            },
            "type": "array",
            "title": "Templates"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "templates"
        ],
        "title": "TrainingTemplateListResponse"
      },
      "TrainingTemplateUpdate": {
        "properties": {
          "default_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Config"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "reset_to_baseline": {
            "type": "boolean",
            "title": "Reset To Baseline",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TrainingTemplateUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}